requirements.txt
pywhatkit==5.4
pyyaml==6.0.1
python-dotenv==1.0.1
requests==2.32.3